draft 관련 메서드 draftController로 이사
