+++ src/main/java/com/example/demo/controller/UsrDraftController.java
@@ -2,6 +2,7 @@ package com.example.demo.controller;
 
 import com.example.demo.service.DraftService;
 import com.example.demo.service.GptService;
+import com.example.demo.service.MemberService;
 import com.example.demo.vo.ResultData;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -21,6 +22,26 @@ public class UsrDraftController {
 
     @Autowired
     private GptService gptService;
+
+    @Autowired
+    private MemberService memberService;
+
+
+    @PostMapping("/verifyGitUser")
+    @ResponseBody
+    public ResultData verifyGitUser(@RequestBody Map<String, String> requestMap) {
+
+        String email = requestMap.get("email");
+        Integer verifiedMemberId = memberService.isVerifiedUser(email);
+
+        if(verifiedMemberId != null) {
+            System.out.println("git email로 찾은 memberID: " + verifiedMemberId);
+            return ResultData.from("S-1", "사용자 인증 완료", "인증된 사용자 id", verifiedMemberId);
+        }else {
+            System.err.println("git email로 찾은 member 없음");
+            return ResultData.from("F-1", "사용자 인증 실패");
+        }
+    }
 
     @PostMapping("/mkRepo")
     @ResponseBody
