+++ src/main/java/com/example/demo/controller/UsrMemberController.java
@@ -198,22 +198,4 @@ public class UsrMemberController {
         );
     }
 
-
-    ////////////////////////////////////////////// CLI ///////////////////////////////////////////////////
-    @PostMapping("/verifyGitUser")
-    @ResponseBody
-    public ResultData verifyGitUser(@RequestBody Map<String, String> requestMap) {
-
-        String email = requestMap.get("email");
-        Integer verifiedMemberId = memberService.isVerifiedUser(email);
-
-        if(verifiedMemberId != null) {
-            System.out.println("git email로 찾은 memberID: " + verifiedMemberId);
-            return ResultData.from("S-1", "사용자 인증 완료", "인증된 사용자 id", verifiedMemberId);
-        }else {
-            System.err.println("git email로 찾은 member 없음");
-            return ResultData.from("F-1", "사용자 인증 실패");
-        }
-    }
-
 }
