+++ src/main/resources/application.yml
@@ -3,7 +3,7 @@ server:
   port: 8080
 
 jwt:
-  secret: 7a36d8d1c41b9e4d8f4260fa38fca973b33c14f2cfd7f1a9d3c848057c12a1e8b2dca3120f2ed014f9a4568ce841fb15b23051cd1e8763b64a0f12c29b2d9d0a
+  secret:
   access-token-expiration-time: 3600000
   refresh-token-expiration-time: 86400000
 
@@ -66,7 +66,7 @@ logging:
 
 sonarqube:
   host: http://localhost:9000
-  token: sqa_ec1e62fbf7d348a0e395e347bb75f0ba16444d61
+  token:
 
 openai:
   api:
